Move analytics scripts into new 'analytics' dir.
