+++ frontend/lib/analytics/analytics.ts
