+++ frontend/lib/analytics/fullstory-globals.d.ts
