+++ frontend/lib/analytics/google-analytics.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import { DetailedHTMLProps, AnchorHTMLAttributes, MouseEvent } from "react";
-import { callOnceWithinMs, getFunctionProperty } from "./util";
-import hardRedirect from './hard-redirect';
+import { callOnceWithinMs, getFunctionProperty } from "../util";
+import hardRedirect from '../hard-redirect';
 
 /**
  * The analytics.js API is provided by Google Analytics:
