+++ frontend/lib/analytics/rollbar-globals.d.ts
