+++ frontend/lib/analytics/tests/analytics.test.ts
