+++ frontend/lib/customer-support-link.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { OutboundLink } from "./google-analytics";
+import { OutboundLink } from "./analytics/google-analytics";
 
 const CUSTOMER_SUPPORT_EMAIL = "support@justfix.nyc";
